Created HashMap factory function
Functions added to HashMap

check if too many buckets were made(checkInLimit.js)

hash(key): takes a key and produces a hash code with it

set(key, value): sets a key with the desired value.
if a key has already been made, the new value will
overwrite the old value

get(key): takes one argument as a key and returns the value that
is assigned to this key. If a key is not found
return null